Ajout de l'affichage des cartes
affichage des cartes, methodes de deck et methodes de card ont été ajouté

+++ HeroLoop/src/entities/Monster.java
@@ -3,7 +3,7 @@ package entities;
 public class Monster extends AbstractEntity {
 	private final double baseStrength;
 
-	public Monster(int hpMax, double hp, double baseStrength, double def, double regen, int damageToAll, int vampirism,
+	public Monster(int hpMax, int hp, double baseStrength, double def, double regen, int damageToAll, int vampirism,
 			int evasivness, String sprite) {
 		
 		super(hpMax, hp, def, regen, damageToAll, vampirism, evasivness, sprite);
